Refactor Work index WorkIndexByResourceBindingID and WorkIndexByClusterResourceBindingID
