+++ pkg/util/helper/binding_test.go
@@ -468,7 +468,7 @@ func TestFindOrphanWorks(t *testing.T) {
 				).WithIndex(
 					&workv1alpha1.Work{},
 					indexregistry.WorkIndexByResourceBindingID,
-					IndexerFuncBasedOnLabel(workv1alpha2.ResourceBindingPermanentIDLabel),
+					indexregistry.GenLabelIndexerFunc(workv1alpha2.ResourceBindingPermanentIDLabel),
 				).Build(),
 				bindingNamespace: "default",
 				bindingName:      "binding",
@@ -517,7 +517,7 @@ func TestFindOrphanWorks(t *testing.T) {
 				).WithIndex(
 					&workv1alpha1.Work{},
 					indexregistry.WorkIndexByResourceBindingID,
-					IndexerFuncBasedOnLabel(workv1alpha2.ResourceBindingPermanentIDLabel),
+					indexregistry.GenLabelIndexerFunc(workv1alpha2.ResourceBindingPermanentIDLabel),
 				).Build(),
 				bindingNamespace: "default",
 				bindingName:      "binding",
@@ -573,7 +573,7 @@ func TestFindOrphanWorks(t *testing.T) {
 				).WithIndex(
 					&workv1alpha1.Work{},
 					indexregistry.WorkIndexByClusterResourceBindingID,
-					IndexerFuncBasedOnLabel(workv1alpha2.ClusterResourceBindingPermanentIDLabel),
+					indexregistry.GenLabelIndexerFunc(workv1alpha2.ClusterResourceBindingPermanentIDLabel),
 				).Build(),
 				bindingNamespace: "",
 				bindingName:      "binding",
@@ -1034,7 +1034,7 @@ func TestDeleteWorkByRBNamespaceAndName(t *testing.T) {
 				c: fake.NewClientBuilder().WithScheme(gclient.NewSchema()).WithIndex(
 					&workv1alpha1.Work{},
 					indexregistry.WorkIndexByResourceBindingID,
-					IndexerFuncBasedOnLabel(workv1alpha2.ResourceBindingPermanentIDLabel),
+					indexregistry.GenLabelIndexerFunc(workv1alpha2.ResourceBindingPermanentIDLabel),
 				).Build(),
 				namespace: "default",
 				name:      "foo",
@@ -1062,7 +1062,7 @@ func TestDeleteWorkByRBNamespaceAndName(t *testing.T) {
 				).WithIndex(
 					&workv1alpha1.Work{},
 					indexregistry.WorkIndexByResourceBindingID,
-					IndexerFuncBasedOnLabel(workv1alpha2.ResourceBindingPermanentIDLabel),
+					indexregistry.GenLabelIndexerFunc(workv1alpha2.ResourceBindingPermanentIDLabel),
 				).Build(),
 				namespace: "default",
 				name:      "foo",
@@ -1089,7 +1089,7 @@ func TestDeleteWorkByRBNamespaceAndName(t *testing.T) {
 				).WithIndex(
 					&workv1alpha1.Work{},
 					indexregistry.WorkIndexByClusterResourceBindingID,
-					IndexerFuncBasedOnLabel(workv1alpha2.ClusterResourceBindingPermanentIDLabel),
+					indexregistry.GenLabelIndexerFunc(workv1alpha2.ClusterResourceBindingPermanentIDLabel),
 				).Build(),
 				name:      "foo",
 				bindingID: "3617252f-b1bb-43b0-98a1-c7de833c472c",
